Add launch file and change to ros parameters from defined const

+++ common.h
@@ -15,11 +15,11 @@ namespace cartographer_ros {
 namespace cartographer_ros_navigation{
   
 // Return distance2 in XY plane
-float Distance2BetweenPoint(const geometry_msgs::Point& point1, const geometry_msgs::Point& point2){
+double Distance2BetweenPoint(const geometry_msgs::Point& point1, const geometry_msgs::Point& point2){
     return (point1.x-point2.x)*(point1.x-point2.x) + (point1.y-point2.y)*(point1.y-point2.y);
 }
     
-float Distance2BetweenPose(const geometry_msgs::Pose& pose1, const geometry_msgs::Pose& pose2){
+double Distance2BetweenPose(const geometry_msgs::Pose& pose1, const geometry_msgs::Pose& pose2){
     return (pose1.position.x-pose2.position.x)*(pose1.position.x-pose2.position.x)+
     (pose1.position.y-pose2.position.y)*(pose1.position.y-pose2.position.y)
     +(pose1.position.z-pose2.position.z)*(pose1.position.z-pose2.position.z);
@@ -33,7 +33,7 @@ geometry_msgs::Point operator+(const geometry_msgs::Point& a, const geometry_msg
     return sum;
 }
 
-geometry_msgs::Point operator*(float a, const geometry_msgs::Point& b){
+geometry_msgs::Point operator*(double a, const geometry_msgs::Point& b){
     geometry_msgs::Point product;
     product.x = a * b.x;
     product.y = a * b.y;
